# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CaptureResultEmailRequest {
  email: string;
  /** If true, send the email now instead of queuing. */
  immediate?: boolean;
  /** Optional shop name to personalise the email. */
  shopName?: string;
}

export interface DispatchResultEmailRequest {
  shopName?: string;
}

export interface ResultEmailStatus {
  ok: boolean;
  sent?: boolean;
  queued?: boolean;
}

/**
 * Full diagnostic result payload.
 */
export type CacheResultRequestPayload = { [key: string]: unknown };

export interface CacheResultRequest {
  shopName?: string;
  /** Full diagnostic result payload. */
  payload: CacheResultRequestPayload;
}

export interface CacheResultStatus {
  ok: boolean;
}

export type CachedResultPayload = { [key: string]: unknown };

export interface CachedResult {
  sid: string;
  shopName?: string | null;
  payload: CachedResultPayload;
  cachedAt: string;
  hasEmail: boolean;
}

export interface ErrorResponse {
  ok: boolean;
  error: string;
}