# lib/api-zod/src/generated/types/cacheResultRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CacheResultRequestPayload } from './cacheResultRequestPayload';

export interface CacheResultRequest {
  shopName?: string;
  /** Full diagnostic result payload. */
  payload: CacheResultRequestPayload;
}