# lib/api-zod/src/generated/types/cachedResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CachedResultPayload } from './cachedResultPayload';

export interface CachedResult {
  sid: string;
  shopName?: string | null;
  payload: CachedResultPayload;
  cachedAt: Date;
  hasEmail: boolean;
}