# lib/api-zod/src/generated/types/captureResultEmailRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CaptureResultEmailRequest {
  email: string;
  /** If true, send the email now instead of queuing. */
  immediate?: boolean;
  /** Optional shop name to personalise the email. */
  shopName?: string;
}